+++ tomcat/src/main/java/org/apache/coyote/common/ContentType.java
@@ -1,4 +1,4 @@
-package org.apache.coyote;
+package org.apache.coyote.common;
 
 import java.util.Arrays;
